// src/app/shared/models/gastos-energ.model.ts
export interface IGastosEnerg {
    id: string;
    desc: string;
    dataAtend: string;
    idade: string;
    sexo: string;
    altura: string;
    peso: string;
    protocolo: string;
    nivelAtiv: string;
    gastoEnergFinal: string;
    classificacao: string;
    massaMagra: string;
}

export interface IGastosEnergMin {
    id: string;
    desc: string;
    dataAtend: string;
}